requirements.txt
# Main Application Frameworks
crewai
streamlit
llama-index
litellm

# Core Tools & Utilities
requests
python-dotenv
PyYAML
sentence-transformers
tavily-python
langchain

# LlamaIndex Integrations
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-vector-stores-weaviate
weaviate-client